Added utils module with common functions for VNET testing (#298)
Moved VnetObjects and VnetAPI from saidahvnet.py and placed it in sai_dash_utils.py
