+++ test/test-cases/functional/ptf/saidashvnet.py
@@ -22,297 +22,12 @@ from unittest import skipIf
 from ptf.testutils import test_param_get
 from sai_base_test import *
 from sai_thrift.sai_headers import *
-
-
-class VNetObjects(SaiHelperSimplified):
-    def setUp(self):
-        super(VNetObjects, self).setUp()
-        self.teardown_objects = list()
-
-    def tearDown(self):
-        super(VNetObjects, self).tearDown()
-
-    def add_teardown_obj(self, func, *args):
-        self.teardown_objects.insert(0, (func, *args))
-
-    def destroy_teardown_obj(self):
-        for obj_func, obj_args in self.teardown_objects:
-            obj_func(*obj_args)
-            self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-
-
-class VNetAPI(VNetObjects):
-    def setUp(self):
-        super(VNetAPI, self).setUp()
-
-    def tearDown(self):
-        self.destroy_teardown_obj()
-        super(VNetAPI, self).tearDown()
-
-    def vip_create(self, vip):
-        """
-        Add VIP for Appliance
-        """
-
-        sai_vip_entry = sai_thrift_vip_entry_t(switch_id=self.switch_id, vip=sai_ipaddress(vip))
-        sai_thrift_create_vip_entry(self.client, sai_vip_entry, action=SAI_VIP_ENTRY_ACTION_ACCEPT)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.vip_remove, sai_vip_entry)
-
-    def vip_remove(self, vip_entry):
-        sai_thrift_remove_vip_entry(self.client, vip_entry)
-
-    def eni_create(self, **kwargs):
-        """
-        Create ENI
-        """
-
-        default_kwargs = {
-            "admin_state": True,
-            "vm_underlay_dip": sai_ipaddress("0.0.0.0"),
-            "vm_vni": 1,
-            "vnet_id": 1,
-            "inbound_v4_stage1_dash_acl_group_id": 0,
-            "inbound_v4_stage2_dash_acl_group_id": 0,
-            "inbound_v4_stage3_dash_acl_group_id": 0,
-            "inbound_v4_stage4_dash_acl_group_id": 0,
-            "inbound_v4_stage5_dash_acl_group_id": 0,
-            "outbound_v4_stage1_dash_acl_group_id": 0,
-            "outbound_v4_stage2_dash_acl_group_id": 0,
-            "outbound_v4_stage3_dash_acl_group_id": 0,
-            "outbound_v4_stage4_dash_acl_group_id": 0,
-            "outbound_v4_stage5_dash_acl_group_id": 0,
-            "inbound_v6_stage1_dash_acl_group_id": 0,
-            "inbound_v6_stage2_dash_acl_group_id": 0,
-            "inbound_v6_stage3_dash_acl_group_id": 0,
-            "inbound_v6_stage4_dash_acl_group_id": 0,
-            "inbound_v6_stage5_dash_acl_group_id": 0,
-            "outbound_v6_stage1_dash_acl_group_id": 0,
-            "outbound_v6_stage2_dash_acl_group_id": 0,
-            "outbound_v6_stage3_dash_acl_group_id": 0,
-            "outbound_v6_stage4_dash_acl_group_id": 0,
-            "outbound_v6_stage5_dash_acl_group_id": 0
-        }
-        default_kwargs.update(kwargs)
-        eni_id = sai_thrift_create_eni(self.client, **default_kwargs)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.assertNotEqual(eni_id, 0)
-        self.add_teardown_obj(self.eni_remove, eni_id)
-
-        return eni_id
-
-    def eni_remove(self, eni_id):
-        sai_thrift_remove_eni(self.client, eni_id)
-
-    def direction_lookup_create(self, vni):
-        """
-        Create direction lookup entry
-        """
-
-        act = SAI_DIRECTION_LOOKUP_ENTRY_ACTION_SET_OUTBOUND_DIRECTION
-
-        direction_lookup_entry = sai_thrift_direction_lookup_entry_t(switch_id=self.switch_id, vni=vni)
-        sai_thrift_create_direction_lookup_entry(self.client, direction_lookup_entry, action=act)
-
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.direction_lookup_remove, direction_lookup_entry)
-
-    def direction_lookup_remove(self, direction_lookup_entry):
-        sai_thrift_remove_direction_lookup_entry(self.client, direction_lookup_entry)
-
-    def eni_mac_map_create(self, eni_id, mac):
-        """
-        Create ENI - MAC mapping
-        """
-
-        eni_ether_address_map_entry = sai_thrift_eni_ether_address_map_entry_t(switch_id=self.switch_id, address=mac)
-        sai_thrift_create_eni_ether_address_map_entry(self.client, eni_ether_address_map_entry, eni_id=eni_id)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.eni_mac_map_remove, eni_ether_address_map_entry)
-
-    def eni_mac_map_remove(self, eni_ether_address_map_entry):
-        sai_thrift_remove_eni_ether_address_map_entry(self.client, eni_ether_address_map_entry)
-
-    def vnet_create(self, vni):
-        """
-        Create VNET
-        """
-
-        vnet_id = sai_thrift_create_vnet(self.client, vni=vni)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.assertNotEqual(vnet_id, 0)
-        self.add_teardown_obj(self.vnet_remove, vnet_id)
-
-        return vnet_id
-
-    def vnet_remove(self, vnet_id):
-        sai_thrift_remove_vnet(self.client, vnet_id)
-
-    def inbound_routing_decap_validate_create(self, eni_id, vni, sip, sip_mask, src_vnet_id):
-        """
-        Create inbound routing entry with
-        SAI_INBOUND_ROUTING_ENTRY_ACTION_VXLAN_DECAP_PA_VALIDATE action
-        """
-
-        inbound_routing_entry = sai_thrift_inbound_routing_entry_t(
-                                        switch_id=self.switch_id, vni=vni,
-                                        eni_id=eni_id, sip=sai_ipaddress(sip),
-                                        sip_mask=sai_ipaddress(sip_mask), priority=0)
-        sai_thrift_create_inbound_routing_entry(self.client, inbound_routing_entry,
-                                                action=SAI_INBOUND_ROUTING_ENTRY_ACTION_VXLAN_DECAP_PA_VALIDATE,
-                                                src_vnet_id=src_vnet_id)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.inbound_routing_remove, inbound_routing_entry)
-
-    def inbound_routing_decap_create(self, eni_id, vni, sip, sip_mask):
-        """
-        Create inbound routing entry with
-        SAI_INBOUND_ROUTING_ENTRY_ACTION_VXLAN_DECAP action
-        """
-
-        inbound_routing_entry = sai_thrift_inbound_routing_entry_t(
-            switch_id=self.switch_id, vni=vni,
-            eni_id=eni_id, sip=sai_ipaddress(sip),
-            sip_mask=sai_ipaddress(sip_mask), priority=0)
-        sai_thrift_create_inbound_routing_entry(self.client, inbound_routing_entry,
-                                                action=SAI_INBOUND_ROUTING_ENTRY_ACTION_VXLAN_DECAP)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.inbound_routing_remove, inbound_routing_entry)
-
-    def inbound_routing_remove(self, inbound_routing_entry):
-        sai_thrift_remove_inbound_routing_entry(self.client, inbound_routing_entry)
-
-    def pa_validation_create(self, sip, vnet_id):
-        """
-        Create source PA validation entry
-        """
-
-        pa_validation_entry = sai_thrift_pa_validation_entry_t(switch_id=self.switch_id,
-                                                               sip=sai_ipaddress(sip),
-                                                               vnet_id=vnet_id)
-        sai_thrift_create_pa_validation_entry(self.client,
-                                              pa_validation_entry,
-                                              action=SAI_PA_VALIDATION_ENTRY_ACTION_PERMIT)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.pa_validation_remove, pa_validation_entry)
-
-    def pa_validation_remove(self, pa_validation_entry):
-        sai_thrift_remove_pa_validation_entry(self.client, pa_validation_entry)
-
-    def outbound_routing_vnet_direct_create(self, eni_id, lpm, dst_vnet_id, overlay_ip):
-        """
-        Create outband vnet direct routing entry
-        """
-
-        outbound_routing_entry = sai_thrift_outbound_routing_entry_t(
-                                switch_id=self.switch_id, eni_id=eni_id,
-                                destination=sai_ipprefix(lpm))
-        sai_thrift_create_outbound_routing_entry(self.client, outbound_routing_entry,
-                                                 dst_vnet_id=dst_vnet_id, overlay_ip=sai_ipaddress(overlay_ip),
-                                                 action=SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET_DIRECT)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.outbound_routing_vnet_direct_remove, outbound_routing_entry)
-
-    def outbound_routing_direct_create(self, eni_id, lpm):
-        """
-        Create outband vnet direct routing entry
-        """
-
-        outbound_routing_entry = sai_thrift_outbound_routing_entry_t(
-                                switch_id=self.switch_id, eni_id=eni_id,
-                                destination=sai_ipprefix(lpm))
-        sai_thrift_create_outbound_routing_entry(self.client, outbound_routing_entry,
-                                                 action=SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_DIRECT)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.outbound_routing_vnet_direct_remove, outbound_routing_entry)
-
-    def outbound_routing_vnet_direct_remove(self, entry):
-        sai_thrift_remove_outbound_routing_entry(self.client, entry)
-
-    def outbound_ca_to_pa_create(self, dst_vnet_id, dip, underlay_dip):
-        """
-        Create outband CA PA mapping
-        """
-
-        ca_to_pa_entry = sai_thrift_outbound_ca_to_pa_entry_t(switch_id=self.switch_id,
-                                                              dst_vnet_id=dst_vnet_id, dip=sai_ipaddress(dip))
-        sai_thrift_create_outbound_ca_to_pa_entry(self.client, ca_to_pa_entry,
-                                                  underlay_dip=sai_ipaddress(underlay_dip), use_dst_vnet_vni=True)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.outbound_ca_to_pa_remove, ca_to_pa_entry)
-
-    def outbound_ca_to_pa_remove(self, ca_to_pa_entry):
-        sai_thrift_remove_outbound_ca_to_pa_entry(self.client, ca_to_pa_entry)
-
-    def router_interface_create(self, port, src_mac=None):
-        """
-        RIF create
-        """
-
-        rif = sai_thrift_create_router_interface(self.client,
-                                                 type=SAI_ROUTER_INTERFACE_TYPE_PORT,
-                                                 virtual_router_id=self.default_vrf,
-                                                 src_mac_address=src_mac, port_id=port)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.router_interface_remove, rif)
-
-        return rif
-
-    def router_interface_remove(self, rif):
-        sai_thrift_remove_router_interface(self.client, rif)
-
-    def nexthop_create(self, rif, ip):
-        """
-        Nexthop create
-        """
-
-        nhop = sai_thrift_create_next_hop(
-            self.client,
-            ip=sai_ipaddress(ip),
-            router_interface_id=rif,
-            type=SAI_NEXT_HOP_TYPE_IP)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.nexthop_remove, nhop)
-
-        return nhop
-
-    def nexthop_remove(self, nhop):
-        sai_thrift_remove_next_hop(self.client, nhop)
-
-    def neighbor_create(self, rif, ip, dmac):
-        """
-        Neighbor create
-        """
-
-        neighbor_entry = sai_thrift_neighbor_entry_t(
-            rif_id=rif, ip_address=sai_ipaddress(ip))
-        sai_thrift_create_neighbor_entry(
-            self.client, neighbor_entry, dst_mac_address=dmac)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.neighbor_remove, neighbor_entry)
-
-    def neighbor_remove(self, entry):
-        sai_thrift_remove_neighbor_entry(self.client, entry)
-
-    def route_create(self, prefix, nhop):
-        """
-        Route create
-        """
-
-        route_entry = sai_thrift_route_entry_t(
-            vr_id=self.default_vrf, destination=sai_ipprefix(prefix))
-        sai_thrift_create_route_entry(
-            self.client, route_entry, next_hop_id=nhop)
-        self.assertEqual(self.status(), SAI_STATUS_SUCCESS)
-        self.add_teardown_obj(self.route_remove, route_entry)
-
-    def route_remove(self, entry):
-        sai_thrift_remove_route_entry(self.client, entry)
+from sai_dash_utils import *
 
 
 @group("draft")
 @disabled  # This is a Demo test. It should not be executed on CI
-class Vnet2VnetCTTest(VNetAPI):
+class Vnet2VnetCTTest(VnetAPI):
     """
     Vnet to Vnet scenario test case Inbound
     """
@@ -351,7 +66,7 @@ class Vnet2VnetCTTest(VNetAPI):
 
 @group("draft")
 @skipIf(test_param_get('bmv2'), "Blocked by Issue #233. Inbound Routing is not supported in BMv2.")
-class Vnet2VnetInboundTest(VNetAPI):
+class Vnet2VnetInboundTest(VnetAPI):
     """
     Inbound Vnet to Vnet scenario test case
     """
@@ -525,7 +240,7 @@ class Vnet2VnetInboundTest(VNetAPI):
 
 @group("draft")
 @skipIf(test_param_get('bmv2'), "Blocked on BMv2 by Issue #236")
-class Vnet2VnetOutboundRouteVnetDirectTest(VNetAPI):
+class Vnet2VnetOutboundRouteVnetDirectTest(VnetAPI):
     """
     Outbound VNet to VNet test scenario with Outbound routing entry
     SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_VNET_DIRECT action
@@ -629,7 +344,7 @@ class Vnet2VnetOutboundRouteVnetDirectTest(VNetAPI):
 
 @group("draft")
 @skipIf(test_param_get('bmv2'), "Blocked on BMv2 by Issue #236")
-class Vnet2VnetOutboundRouteDirectTest(VNetAPI):
+class Vnet2VnetOutboundRouteDirectTest(VnetAPI):
     """
     Outbound VNet to VNet test scenario with Outbound routing entry
     SAI_OUTBOUND_ROUTING_ENTRY_ACTION_ROUTE_DIRECT action
@@ -720,7 +435,7 @@ class Vnet2VnetOutboundRouteDirectTest(VNetAPI):
 
 @group("draft")
 @skipIf(test_param_get('bmv2'), "Blocked on BMv2 by Issue #236")
-class VnetRouteTest(VNetAPI):
+class VnetRouteTest(VnetAPI):
     """
     Vnet to Vnet scenario test case Outbound
     """
